lower/mod.rs hands its widening/extern/strict-mode tail to mod_d.rs, a pure text move under the 800-line ceiling
