+++ crates/almide-mir/src/lower/mod.rs
@@ -441,427 +441,8 @@ fn extern_wasm_abi(ty: &Ty) -> Option<crate::WasmAbi> {
     }
 }
 
-/// A sized-int WIDENING conversion call (`int8.to_int64(x)`, `uint32.to_int64(x)`, …)
-/// whose runtime is the IDENTITY on the canonical-i64 slot value: every integer width
-/// lives sign-/zero-extended in one i64 (the `Ty` docs + `extern_wasm_abi` pin this),
-/// and the Rust runtime is `n as i64` over that already-canonical value (`u64 as i64`
-/// is the same bit-reinterpret the slot already holds). Returns the operand expr when
-/// the shape applies — the lowering forwards the operand's value with NO call, and
-/// `count_ir_calls` skips the node by the SAME predicate (mir == ir by construction).
-pub fn identity_int_widening_call(e: &IrExpr) -> Option<&IrExpr> {
-    let IrExprKind::Call {
-        target: CallTarget::Module { module, func, .. },
-        args,
-        ..
-    } = &e.kind
-    else {
-        return None;
-    };
-    is_identity_int_widening(module.as_str(), func.as_str(), args).then(|| &args[0])
-}
-
-/// The (module, func, args) core of [`identity_int_widening_call`], for the
-/// producers that hold an UNPACKED Module call (`lower_pure_module_value_call`)
-/// — one predicate, so the elision and the caps counter cannot drift apart.
-pub fn is_identity_int_widening(module: &str, func: &str, args: &[IrExpr]) -> bool {
-    if args.len() != 1 || func != "to_int64" {
-        return false;
-    }
-    if !matches!(
-        module,
-        "int" | "int8" | "int16" | "int32" | "int64" | "uint8" | "uint16" | "uint32" | "uint64"
-    ) {
-        return false;
-    }
-    matches!(
-        args[0].ty,
-        Ty::Int
-            | Ty::Int8
-            | Ty::Int16
-            | Ty::Int32
-            | Ty::Int64
-            | Ty::UInt8
-            | Ty::UInt16
-            | Ty::UInt32
-            | Ty::UInt64
-    )
-}
-
-/// A `float.from_int(x)` call over an `Int` — the sitofp floor (#806 step 2):
-/// the lowering emits ONE `PrimKind::F64FromInt` (a `f64.convert_i64_s` in the
-/// render, `as f64` natively) instead of the self-host runtime CALL, and
-/// `count_ir_calls` skips the node by this SAME predicate (`mir == ir` by
-/// construction). Returns the operand expr when the shape applies.
-pub fn float_from_int_prim_call(e: &IrExpr) -> Option<&IrExpr> {
-    let IrExprKind::Call {
-        target: CallTarget::Module { module, func, .. },
-        args,
-        ..
-    } = &e.kind
-    else {
-        return None;
-    };
-    is_float_from_int_prim(module.as_str(), func.as_str(), args).then(|| &args[0])
-}
-
-/// The (module, func, args) core of [`float_from_int_prim_call`] — same
-/// single-predicate discipline as [`is_identity_int_widening`].
-pub fn is_float_from_int_prim(module: &str, func: &str, args: &[IrExpr]) -> bool {
-    module == "float" && func == "from_int" && args.len() == 1 && matches!(args[0].ty, Ty::Int)
-}
-
-/// The `@extern(wasm, module, name)` attribute on a function, iff present (the
-/// browser-import case — a `rust`/`rs` target keeps walling: there is no wasm host
-/// for it, so emitting an import would be a hollow lie). Returns `(module, name)`.
-fn extern_wasm_target(func: &IrFunction) -> Option<(String, String)> {
-    func.extern_attrs.iter().find_map(|a| {
-        if a.target.as_str() == "wasm" {
-            Some((
-                a.module.as_str().to_string(),
-                a.function.as_str().to_string(),
-            ))
-        } else {
-            None
-        }
-    })
-}
-
-/// Lower a body-less `@extern(wasm, module, name)` function to a thin wasm-IMPORT
-/// call body (the browser dom/fetch/timer/console host stubs). The function becomes
-/// a `(call $__import_module_name <params>)` that returns the host's result —
-/// FAITHFUL: its behavior IS the host's, so it calls the host, it does NOT fabricate
-/// a value (an `Opaque`/`0` would be a silent miscompile). The wasm module is valid;
-/// a browser host satisfies the import (it does not instantiate under wasmtime, which
-/// is expected — these fns are 🟡 lower, not byte-matchable on the wasmtime oracle).
-///
-/// Returns `Ok(Some(MirFunction))` when this is a wasm-extern fn whose param + return
-/// types all map to flat valtypes; `Ok(None)` when it is NOT a wasm-extern (the caller
-/// lowers it normally); `Err(Unsupported)` when a param/return type has no flat-valtype
-/// ABI (WALL — never guess a signature). SOUNDNESS: a `rust`/`rs` extern is NOT a wasm
-/// import (no wasm host) → `extern_wasm_target` is `None` → it keeps walling.
-fn try_lower_extern_wasm(func: &IrFunction) -> Result<Option<MirFunction>, LowerError> {
-    let Some((module, name)) = extern_wasm_target(func) else {
-        return Ok(None);
-    };
-    // Bind params to fresh MIR values (the borrow-by-default convention) — a heap param
-    // is a borrowed i32 pointer, a scalar an i64 local; both are read into the call.
-    let mut ctx = LowerCtx {
-        fn_name: func.name.as_str().to_string(),
-        ..Default::default()
-    };
-    let params = ctx.bind_params(&func.params)?;
-    // The import-call args + their per-arg valtypes, parallel to the params. A heap param
-    // is BORROWED (a `Handle` — the caller owns it, no refcount change here); a scalar is
-    // passed by value (`Scalar`). The ABI of each comes from the DECLARED param type.
-    let mut args: Vec<crate::CallArg> = Vec::new();
-    let mut abi: Vec<crate::WasmAbi> = Vec::new();
-    for (p, ip) in params.iter().zip(func.params.iter()) {
-        let a = extern_wasm_abi(&ip.ty).ok_or_else(|| {
-            LowerError::Unsupported(format!(
-                "@extern(wasm) param type {:?} has no flat wasm valtype (not lowered to an import)",
-                ip.ty
-            ))
-        })?;
-        abi.push(a);
-        args.push(if p.repr.is_heap() {
-            crate::CallArg::Handle(p.value)
-        } else {
-            crate::CallArg::Scalar(p.value)
-        });
-    }
-    // The result: `Unit` → a void import (no MIR result); else map the return type to its
-    // valtype + a fresh dst the call binds. A heap return is a FRESH OWNED pointer the host
-    // returns (the caller now owns it — moved out as `ret`, like an `Alloc` result).
-    let (dst, result, result_abi, ret) = if matches!(func.ret_ty, Ty::Unit) {
-        (None, None, None, None)
-    } else {
-        let rabi = extern_wasm_abi(&func.ret_ty).ok_or_else(|| {
-            LowerError::Unsupported(format!(
-                "@extern(wasm) return type {:?} has no flat wasm valtype (not lowered to an import)",
-                func.ret_ty
-            ))
-        })?;
-        let repr = repr_of(&func.ret_ty)?;
-        let d = ctx.fresh_value();
-        (Some(d), Some(repr), Some(rabi), Some(d))
-    };
-    ctx.ops.push(Op::CallImport {
-        dst,
-        module,
-        name,
-        args,
-        abi,
-        result,
-        result_abi,
-    });
-    Ok(Some(MirFunction {
-        name: func.name.as_str().to_string(),
-        params,
-        ops: ctx.ops,
-        // A wasm import reaches a BROWSER host capability (dom/fetch/timer/console), which is
-        // OUTSIDE the v1 WASI-floor cap vocabulary (Stdout/Entropy/CliArgs/FsRead). So it
-        // declares no MODELED cap here; the `CallImport` reaches no modeled WASI cap either,
-        // so `used ⊆ declared` holds vacuously — honest (it is not claimed to reach a WASI cap).
-        ret,
-        declared_caps: Vec::new(),
-        heap_slot_masks: Default::default(),
-    }))
-}
-
-/// Lower one function to MIR. Parameters are seeded first (the v1 borrow-by-
-/// default calling convention — see [`LowerCtx::bind_params`]), then the body.
-/// STRICT VALUE MODE (flight-evidence-gaps F2 — retiring the deferred-Const):
-/// when set, every lowering site that would fall back to `Op::Const` (the
-/// deferred ZERO whose only legitimate consumer is the caps-counting
-/// classifier) REFUSES instead — a walled function can never print a silently
-/// wrong value (the `prim.handle(<literal>)` → address-0 class). Set by the
-/// render/output entrypoints (render_program); the classifier and the in-process
-/// unit tests keep the permissive caps-counting behavior.
-///
-/// THREAD-LOCAL, for the same reason [`mod_p2::NEVER_ERR_LIFTED_FNS`] is:
-/// lowering runs single-threaded per program, so per-thread state carries the
-/// mode faithfully on the real render path. A process-global `AtomicBool` did
-/// not: `cargo test` runs a crate's unit tests as many threads in ONE process,
-/// so a test that enabled strict mode leaked it into every test that ran after
-/// it on any thread. `lower::tests::scalar_loop_with_break_walls` asserts a
-/// break/continue wall and intermittently saw a STRICT-mode wall instead —
-/// order-dependent, invisible when the test ran alone, and the doc-comment above
-/// claimed the opposite ("the in-process unit tests keep the permissive
-/// behavior") while nothing enforced it.
-pub fn set_strict_values(on: bool) {
-    STRICT_VALUES.with(|s| s.set(on));
-}
-
-/// Enable strict mode for a scope and restore the previous value on drop.
-///
-/// Prefer this over the bare setter. Thread-locality alone is not enough: it
-/// isolates the PARALLEL test runner (libtest gives each test its own thread),
-/// but `--test-threads=1` runs every test on the same thread, where an
-/// unrestored setter leaks exactly as the process-global did. It also matters
-/// outside tests — a long-lived process that compiles more than one program
-/// (the playground, the language server) would otherwise have the first strict
-/// render pin the mode for everything after it.
-#[must_use = "the guard restores the previous mode when dropped; binding it to `_` restores immediately"]
-pub struct StrictValuesGuard(bool);
-
-impl StrictValuesGuard {
-    pub fn set(on: bool) -> Self {
-        let prev = strict_values();
-        set_strict_values(on);
-        Self(prev)
-    }
-}
-
-impl Drop for StrictValuesGuard {
-    fn drop(&mut self) {
-        set_strict_values(self.0);
-    }
-}
-
-thread_local! {
-    static STRICT_VALUES: std::cell::Cell<bool> = const { std::cell::Cell::new(false) };
-    /// The DESUGAR VAR BAND cursor — see [`desugar_var_seed`].
-    static DESUGAR_VAR_CURSOR: std::cell::Cell<u32> = const { std::cell::Cell::new(DESUGAR_VAR_BAND) };
-}
-
-/// The floor of the synthetic-variable band the desugar/lowering rewrites mint
-/// from. Frontend `VarId`s are dense from 0 and never approach this;
-/// `fresh_synth_var` counts DOWN from `u32::MAX` (the other reserved band).
-pub(crate) const DESUGAR_VAR_BAND: u32 = 0x4000_0000;
-
-/// Reset the band cursor — called once per function at the desugar-chain entry
-/// (`apply_pre_lower_desugars`), so every desugar of the same body is
-/// DETERMINISTIC (the counting pass and the lowering pass re-derive identical
-/// trees — the desugar-before-both discipline) and the band is per-function
-/// bounded.
-pub(crate) fn reset_desugar_var_band() {
-    DESUGAR_VAR_CURSOR.with(|c| c.set(DESUGAR_VAR_BAND));
-}
-
-/// The `Op::Return` readiness probe (return-op-eradication R2, the #1418
-/// recipe): `ALMIDE_BANG_RETURN=1` turns OFF the per-position `!` desugars in
-/// BOTH the counted tree and the lowering (the desugar-before-both discipline
-/// holds — an un-desugared `Unwrap` adds no calls), so every `!` reaches the
-/// bind-position rule or WALLS loudly. The walls are the decline matrix; the
-/// deletion (R3) fires when it is empty.
-pub(crate) fn bang_return_probe() -> bool {
-    static PROBE: std::sync::OnceLock<bool> = std::sync::OnceLock::new();
-    *PROBE.get_or_init(|| std::env::var_os("ALMIDE_BANG_RETURN").is_some())
-}
-
-/// Reserve a fresh 4096-id chunk in the synthetic band and return its first id.
-///
-/// THE fresh-variable discipline for every IR→IR rewrite (the reference-compiler
-/// rule: fresh ids come from a MONOTONIC allocator owned by the context — never
-/// from scanning a subtree for `max + 1`). The `max_var_id(subtree) + k` scheme
-/// this replaces minted from whatever ids the measured SUBTREE happened to
-/// contain, so a pass that introduced pattern binds ABOVE another pass's subtree
-/// made the two mint the SAME id for two different-typed variables — one render
-/// local, two types, invalid wasm (`Ok(p1: Int)` colliding with a nested
-/// `err(e1: String)` bind, found via the ?? → match desugar experiment,
-/// 2026-08-15). Chunks are disjoint by construction, so call order never
-/// matters; passes that allocate sequentially keep doing `seed, seed+1, …`
-/// inside their own chunk.
-pub(crate) fn desugar_var_seed() -> u32 {
-    DESUGAR_VAR_CURSOR.with(|c| {
-        let seed = c.get();
-        // Well below fresh_synth_var's descending band even if a pathological
-        // function burns thousands of chunks.
-        debug_assert!(seed < u32::MAX / 2, "desugar var band exhausted");
-        c.set(seed + 4096);
-        seed
-    })
-}
-
-pub(crate) fn strict_values() -> bool {
-    STRICT_VALUES.with(|s| s.get())
-}
-
-pub(crate) fn strict_const_wall(what: &str) -> LowerError {
-    LowerError::Unsupported(format!(
-        "scalar {what} outside the value subset cannot be faithfully computed in this          brick (the permissive caps-counting path defers it to Const 0; STRICT value          mode refuses instead of risking a silently wrong value)"
-    ))
-}
-
-pub fn lower_function(
-    func: &IrFunction,
-    globals: &HashMap<VarId, Ty>,
-) -> Result<MirFunction, LowerError> {
-    // The main function only; any lambda-lifted auxiliaries are dropped (callers that
-    // need them — render/verify paths — use `lower_function_all`). Sound while no lambda
-    // lifting is wired (lifted is empty); when it is, those paths verify the auxiliaries.
-    let mut all = lower_function_all(func, globals)?;
-    Ok(all.remove(0))
-}
-
-/// Lower a function to its MIR plus any lambda-lifted auxiliary functions (index 0 is the
-/// main function). The closures machinery lifts `let f = (x) => …` bodies into fresh
-/// functions accumulated in `LowerCtx::lifted`; this returns them so the program assembler
-/// can table + verify them. With no lifting wired the result is just `[main]`.
-pub fn lower_function_all(
-    func: &IrFunction,
-    globals: &HashMap<VarId, Ty>,
-) -> Result<Vec<MirFunction>, LowerError> {
-    lower_function_all_with_types(func, globals, &RecordLayouts::new())
-}
-
-/// Substitute every `Var { id: from }` in `e` with `Var { id: to }` — the binder rebind
-/// the defunc match-arm transforms use (`some(b) => X` becomes `X[b := payload_var]`).
-pub(crate) fn subst_var_ir(e: &almide_ir::IrExpr, from: VarId, to: VarId) -> almide_ir::IrExpr {
-    fn walk(e: almide_ir::IrExpr, from: VarId, to: VarId) -> almide_ir::IrExpr {
-        let mut e = e.map_children(&mut |c| walk(c, from, to));
-        if let almide_ir::IrExprKind::Var { id } = &mut e.kind {
-            if *id == from {
-                *id = to;
-            }
-        }
-        e
-    }
-    walk(e.clone(), from, to)
-}
-
-/// Resolve a TYPE NAME against the record registry, accepting the BARE spelling of a
-/// cross-module type when it is UNAMBIGUOUS: the frontend qualifies an imported DECL
-/// (`types_mod.Lin`) but leaves some USE-site `Ty::Named`s bare (`Lin` — the alias-typed
-/// annotation `tm.Lin` resolves to the decl's own Sym), so an exact miss falls back to
-/// the unique `".{name}"`-suffixed key. Two modules exporting the same bare name stay
-/// unresolved (`None`) — the caller walls, never a wrong-layout guess. Returns the
-/// CANONICAL registry key, which is also the drop-fn identity (`$__drop_<canonical>`),
-/// so lowering-side routing and the decl-side generators can never disagree on a name.
-pub(crate) fn canonical_record_key<'a>(layouts: &'a RecordLayouts, name: &str) -> Option<&'a str> {
-    if let Some((k, _)) = layouts.get_key_value(name) {
-        return Some(k.as_str());
-    }
-    let suffix = format!(".{name}");
-    let mut found: Option<&'a str> = None;
-    for k in layouts.keys() {
-        if k.ends_with(&suffix) {
-            if found.is_some() {
-                return None; // ambiguous bare name — walled, never a guess
-            }
-            found = Some(k.as_str());
-        }
-    }
-    found
-}
-
-/// The [`canonical_record_key`] resolution over a NAME SET (the drop generators'
-/// `rec_names`) instead of the layout map — the same exact-then-unique-suffix rule.
-pub(crate) fn canonical_name_in<'a>(
-    names: &'a std::collections::HashSet<String>,
-    name: &str,
-) -> Option<&'a str> {
-    if let Some(k) = names.get(name) {
-        return Some(k.as_str());
-    }
-    let suffix = format!(".{name}");
-    let mut found: Option<&'a str> = None;
-    for k in names {
-        if k.ends_with(&suffix) {
-            if found.is_some() {
-                return None;
-            }
-            found = Some(k.as_str());
-        }
-    }
-    found
-}
-
-/// Build the [`RecordLayouts`] registry from a program's type declarations — the
-/// VALUE-MODEL field structure the lowering consults to materialize records and
-/// resolve `r.x`. Each `type R = { … }` becomes `R → (generic params, fields)`;
-/// variant / alias decls carry no flat record layout and are skipped (a record
-/// VARIANT is a separate, tagged shape — out of this brick). Call once per
-/// program and pass the result into [`lower_function_all_with_types`].
-pub fn build_record_layouts(type_decls: &[almide_ir::IrTypeDecl]) -> RecordLayouts {
-    let mut out = RecordLayouts::new();
-    for decl in type_decls {
-        if let almide_ir::IrTypeDeclKind::Record { fields } = &decl.kind {
-            let generics = decl
-                .generics
-                .as_ref()
-                .map(|gs| gs.iter().map(|g| g.name).collect())
-                .unwrap_or_default();
-            let field_tys = fields.iter().map(|f| (f.name, f.ty.clone())).collect();
-            out.insert(decl.name.as_str().to_string(), (generics, field_tys));
-        }
-    }
-    out
-}
-
-/// Build the [`VariantLayouts`] registry from a program's type declarations — the
-/// VALUE-MODEL tag + per-constructor field structure the ADT bricks consult to construct,
-/// `match`, and drop a custom variant. Each `type V = A(..) | B { .. } | C` becomes
-/// `V → VariantLayout { tag-indexed cases, slot_count }`; record / alias decls carry no
-/// variant layout and are skipped. The tag is the declaration index and tuple-constructor
-/// fields are named `_0`, `_1`, … — both matching v0's `emit_wasm` registration, so the
-/// backends agree on tag and field identity. Call once per program and pass the result
-/// into [`lower_function_all_with_layouts`].
-/// Does `e` contain ANY call node (Named/Module/Method/Computed Call, RuntimeCall,
-/// TailCall)? Used to gate synthesized-expr admissions (a default-field fill) whose calls
-/// the counted IR would not see (the caps `mir == ir` invariant).
-pub fn expr_contains_call(e: &almide_ir::IrExpr) -> bool {
-    use almide_ir::visit::{walk_expr, IrVisitor};
-    struct C(bool);
-    impl IrVisitor for C {
-        fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
-            if matches!(
-                e.kind,
-                almide_ir::IrExprKind::Call { .. }
-                    | almide_ir::IrExprKind::RuntimeCall { .. }
-                    | almide_ir::IrExprKind::TailCall { .. }
-            ) {
-                self.0 = true;
-            }
-            walk_expr(self, e);
-        }
-    }
-    let mut c = C(false);
-    almide_ir::visit::IrVisitor::visit_expr(&mut c, e);
-    c.0
-}
-
 include!("mod_b.rs");
 include!("mod_c.rs");
+include!("mod_d.rs");
 include!("crossmod_toplet_bridge.rs");
 include!("continuation_lift.rs");
